refactor: move layers outside of map component
